Crate staticvec

Note: the complete list of things not available when using default-features = false for #![no_std] compatibility is as follows:

[StaticVec::sorted]

[StaticVec::into_vec] (and the corresponding Into impl)

[StaticVec::from_vec] (and the corresponding From impl)

the implementation of the Read trait for StaticVec

the implementation of the BufRead trait for StaticVec

the implementation of the io::Write trait for StaticVec

the implementation of From for StaticString from String

the implementations of PartialEq and PartialOrd against String for StaticString

the implementation of Error for StringError

the bounds_to_string function unique to this crate and implemented by several of the iterators in it

Macros

sortedstaticvec
Accepts an array of any primitive Copy type that has a PartialOrd implementation, sorts it, and creates a new StaticVec instance from the result in a fully const context compatible manner.

staticstring
Creates a new StaticString from an &str literal. This macro can be used in const contexts, in keeping with the other ones in this crate.

staticvec
Creates a new StaticVec from a vec!-style pseudo-slice. The newly created StaticVec will have a capacity and length exactly equal to the number of elements in the so-called slice. The “array-like” [value; N] syntax is also supported, and both forms can be used in const contexts. This macro has no particular trait impl requirements for the input type.

Structs

CapacityError
This error indicates that an operation was attempted that would have increased the length value of a StaticVec, but the StaticVec was already at its maximum capacity of N.

PushCapacityError
This error indicates that a push was attempted into a StaticVec which failed because the StaticVec was already at maximum capacity. It contains the value that failed to be pushed so that it can be reused if needed.

StaticHeap
A priority queue implemented as a binary heap, built around an instance of StaticVec<T, N>.

StaticHeapDrainSorted
A sorted “draining” iterator over the elements of a StaticHeap.

StaticHeapIntoIterSorted
A sorted “consuming” iterator over the elements of a StaticHeap.

StaticHeapPeekMut
A struct wrapping a mutable reference to the greatest (or “maximal”) item in a StaticHeap.

StaticString
A fixed-capacity String-like struct built around an instance of StaticVec<u8, N>.

StaticVec
A Vec-like struct (mostly directly API-compatible where it can be) implemented with const generics around an array of fixed N capacity.

StaticVecDrain
A “draining” iterator, analogous to vec::Drain. Instances of StaticVecDrain are created by the drain_iter method on StaticVec, as while the drain method does have a similar purpose, it works by immediately returning a new StaticVec as opposed to an iterator.

StaticVecIntoIter
A “consuming” iterator that reads each element out of a source StaticVec by value.

StaticVecIterConst
Similar to Iter, but specifically implemented with StaticVecs in mind.

StaticVecIterMut
Similar to IterMut, but specifically implemented with StaticVecs in mind.

StaticVecSplice
A “splicing” iterator, analogous to vec::Splice. Instances of StaticVecSplice are created by the splice method on StaticVec.

Enums

StringError
This enum represents several different possible “error states” that may be encountered while using a StaticString.